// mantalents/src/interfaces/fetchPersonas.tsx
// Generated by https://quicktype.io

export interface FetchPersonas {
    total:    number;
    personas: Persona[];
}

export interface Persona {
    _id:               string;
    identificacion:    string;
    nombre:            string;
    usuario?:           Usuario;
    correoelectronico: string;
    direccion:         string;
    telefono:          string;
}

export interface Usuario {
    _id:    string;
    nombre: string;
}